Change description for column related tests.
THe column releated tests that allow a schema require a description.
This was hard coded to 'pgtapgen'. This is useless in tests results
and makes finding the problem very difficult.

The description is now the schema.table.column_name with an suffix
to hint at which particular test it might be.