Restsore extenstion creation to pg_tapgen.
But add `--create-extension` and `--no-create-extension` options to
allow it to be explicitly included or excluded in the output.

Also, fix a bug where pg_tapgen died with an error if no database options were
specified. It should be fine to let them all fall back on the usual libpq
defaults. And now it is.