Refactor pg_prove to subclass App::Prove.
This is more in line with how it *should* be written. The impetus for this was
to address failed option parsing when bundling options, such as `pg_prove -vd
foo`. That really comes down to a bug in Getopt::Long
(https://rt.cpan.org/Ticket/Display.html?id=114335), but in coming up with a
workaround for that, it made sense to properly subclass App::Prove and better
let it do its thing. That workaround, though, boy howdy. We have to
monkey-patch App::Prove to replace GetOptions(). But we just do that for the
duration of the call to process_args() (from the subclass implementation), so
the exposure is limited.